Remove unnecessary initializations
You set all of these in the Grid __init__ function anyway, so you
don't need to declare them at the class scope.

Additionally, there's a gotcha when initializing at the class scope:
instances get copies. For reference types, like arrays, this isn't a
deep copy, so there's an aliasing risk. Not a problem in this case,
but it sets off alarm bells.